More folder restructuring for Storybook sidebar.
