+++ app/javascript/crayons/navigation/NavigationTabs/__stories__/navigationTab.html.stories.jsx
@@ -4,15 +4,6 @@ import '../../../storybook-utiltiies/designSystem.scss';
 
 export default { title: 'Components/Navigation/Tabs/HTML' };
 
-export const Description = () => (
-  <div className="container">
-    <h2>Navigation: Tabs</h2>
-    <p>Use tabs as 2nd level navigation or filtering options.</p>
-  </div>
-);
-
-Description.story = { name: 'description' };
-
 export const Default = () => (
   <div className="crayons-tabs">
     <a href="/" className="crayons-tabs__item crayons-tabs__item--current">
